+++ lib/folio/engine.rb
@@ -5,10 +5,19 @@ module Folio
     isolate_namespace Folio
 
     config.to_prepare do
-      Devise::SessionsController.layout 'folio/console/devise'
-      Devise::ConfirmationsController.layout 'folio/console/devise'
-      Devise::UnlocksController.layout 'folio/console/devise'
-      Devise::PasswordsController.layout 'folio/console/devise'
+      [
+        Devise::ConfirmationsController,
+        Devise::OmniauthCallbacksController,
+        Devise::PasswordsController,
+        Devise::RegistrationsController,
+        Devise::SessionsController,
+        Devise::UnlocksController,
+
+        Devise::InvitationsController,
+        DeviseInvitable::RegistrationsController,
+      ].each do |controller|
+        controller.send(:include, Folio::DeviseExtension)
+      end
 
       Dir.glob(Rails.root + 'app/decorators/**/*_decorator*.rb').each do |c|
         require_dependency(c)